Wiring specific version of ES rule explicitly.
