+++ full-backend-tests/src/test/java/org/graylog/storage/elasticsearch6/ElasticsearchInstanceES6Factory.java
@@ -24,6 +24,6 @@ import org.testcontainers.containers.Network;
 public class ElasticsearchInstanceES6Factory implements ElasticsearchInstanceFactory {
     @Override
     public ElasticsearchInstance create(Network network) {
-        return new ElasticsearchInstanceES6(network);
+        return ElasticsearchInstanceES6.create(network);
     }
 }
