+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/CountsES6IT.java
@@ -1,5 +1,6 @@
 package org.graylog.storage.elasticsearch6;
 
+import org.graylog.storage.elasticsearch6.testing.ElasticsearchInstanceES6;
 import org.graylog.testing.elasticsearch.ElasticsearchInstance;
 import org.graylog2.indexer.counts.CountsAdapter;
 import org.graylog2.indexer.counts.CountsIT;
@@ -7,7 +8,7 @@ import org.junit.Rule;
 
 public class CountsES6IT extends CountsIT {
     @Rule
-    public final ElasticsearchInstance elasticsearch = ElasticsearchInstance.create();
+    public final ElasticsearchInstance elasticsearch = ElasticsearchInstanceES6.create();
 
     @Override
     protected ElasticsearchInstance elasticsearch() {
