+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/IndicesES6IT.java
@@ -1,5 +1,6 @@
 package org.graylog.storage.elasticsearch6;
 
+import org.graylog.storage.elasticsearch6.testing.ElasticsearchInstanceES6;
 import org.graylog.testing.elasticsearch.ElasticsearchInstance;
 import org.graylog2.indexer.indices.IndicesAdapter;
 import org.graylog2.indexer.indices.IndicesIT;
@@ -8,7 +9,7 @@ import org.junit.Rule;
 
 public class IndicesES6IT extends IndicesIT {
     @Rule
-    public final ElasticsearchInstance elasticsearch = ElasticsearchInstance.create();
+    public final ElasticsearchInstance elasticsearch = ElasticsearchInstanceES6.create();
 
     @Override
     protected ElasticsearchInstance elasticsearch() {
